requirements.txt
annotated-types==0.6.0
asgiref==3.7.2
beautifulsoup4==4.12.2
Django==4.2.8
django-bootstrap-v5==1.0.11
django-htmx==1.17.2
django-ninja==1.1.0
humanize==4.9.0
Jinja2==3.1.2
MarkupSafe==2.1.3
pydantic==2.5.3
pydantic_core==2.14.6
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.9.0
